Revert "Added some testing, for some reason ULONG_MAX was not giving the max unsigned long, UINT_MAX is giving the max for unsigned long, sounds crazy, but it's true according to climits docs"
This reverts commit a83638811c228993d1b51943595f0f123f0e3c93.

+++ src/NodeData.cpp
@@ -104,7 +104,7 @@ adak_key NodeData::findEndKey(int creationRate){
 }
 
 int NodeData::getMinKeyIndex(std::vector<Keyspace*> keyspaces){
-    unsigned long min = UINT_MAX;
+    unsigned long min = ULONG_MAX;
     int index = -1;
     for(int i =0; i < keyspaces.size(); i++){
         if(keyspaces[i]->getStart() < min && keyspaces[i]->isKeyAvailable()){
